Core: IsNaN(Complex)
git-svn-id: svn://ultimatepp.org/upp/trunk@13643 f0d560ea-af0d-0410-9eb7-867de7ffcac7

+++ uppsrc/Core/JSON.cpp
@@ -140,7 +140,7 @@ String AsJSON(const Value& v, const String& sep, bool pretty)
 		r << "]";
 		return r;
 	}
-	if(IsNumber(v) && (IsNull(v) || IsNaN(v)))
+	if(IsNumber(v) && (IsNull(v) || IsNaN((double)v)))
 		return "null";
 	if(v.GetType() == INT_V)
 		return Format("%d", (int)v);
